+++ src/main/java/com/chuseok22/timemateserver/meeting/application/controller/AvailabilityTimeController.java
@@ -1,7 +1,6 @@
 package com.chuseok22.timemateserver.meeting.application.controller;
 
 import com.chuseok22.timemateserver.common.application.aop.LogMonitoringInvocation;
-import com.chuseok22.timemateserver.meeting.application.controller.docs.AvailabilityTimeControllerDocs;
 import com.chuseok22.timemateserver.meeting.application.dto.request.UpsertAvailabilityRequest;
 import com.chuseok22.timemateserver.meeting.core.service.AvailabilityTimeService;
 import jakarta.validation.Valid;
