+++ src/main/java/com/chuseok22/timemateserver/meeting/application/controller/ParticipantController.java
@@ -1,7 +1,6 @@
 package com.chuseok22.timemateserver.meeting.application.controller;
 
 import com.chuseok22.timemateserver.common.application.aop.LogMonitoringInvocation;
-import com.chuseok22.timemateserver.meeting.application.controller.docs.ParticipantControllerDocs;
 import com.chuseok22.timemateserver.meeting.application.dto.request.ParticipantLoginRequest;
 import com.chuseok22.timemateserver.meeting.application.dto.response.ParticipantInfoResponse;
 import com.chuseok22.timemateserver.meeting.core.service.ParticipantService;
